requirements.txt
customtkinter==5.2.1
requests==2.31.0
pytz==2024.1
pyttsx3==2.90
pillow==10.1.0
ollama==0.1.5
python-dotenv==1.0.0
tkcalendar==1.6.1 